requirements.txt
django==4.0
psycopg2_binary==2.9.2
djangorestframework==3.13.1
django-filters==21.1
pillow==10.1.0
djangorestframework-simplejwt==5.3.0 
pyjwt==1.7.1